+++ app/src/main/java/ar/nablab/mimerendero/repositories/RepositoriesModule.kt
@@ -1,0 +1,16 @@
+package ar.nablab.mimerendero.repositories
+
+import dagger.Binds
+import dagger.Module
+import dagger.hilt.InstallIn
+import dagger.hilt.android.components.ViewModelComponent
+import dagger.hilt.android.scopes.ViewModelScoped
+
+@InstallIn(ViewModelComponent::class)
+@Module
+interface RepositoriesModule {
+
+    @ViewModelScoped
+    @Binds
+    fun bindMapRepository(mapRepositoryImpl: MapRepositoryImpl): MapRepository
+}