+++ CHANGELOG.md
@@ -1,3 +1,11 @@
+# v0.1.0
+## Added
+- Place data model
+- Google Maps connection
+- Firebase connection
+- Write and read from FirebaseDB implementation
+- Detekt code analysis tool
+
 # v0.0.0
 ## Added
 - Scaffolding